backend/requirements.txt
Flask
Flask-Cors
pymongo
psycopg2-binary
sentence-transformers
torch
scikit-learn
celery
redis
